# core/vm: Stack underflow/overflow check optimization