--- a/core/vm/interpreter.go
+++ b/core/vm/interpreter.go
@@ -187,10 +187,8 @@ func (in *EVMInterpreter) Run(contract *Contract, input []byte, readOnly bool) (
 		operation := in.cfg.JumpTable[op]
 		cost = operation.constantGas // For tracing
 		// Validate stack
-		if sLen := stack.len(); sLen < operation.minStack {
-			return nil, &ErrStackUnderflow{stackLen: sLen, required: operation.minStack}
-		} else if sLen > operation.maxStack {
-			return nil, &ErrStackOverflow{stackLen: sLen, limit: operation.maxStack}
+		if !stack.between(operation.minStack, operation.maxStack) {
+			return nil, &ErrStackUnderOverflow{stackLen: stack.len(), minStack: operation.minStack, maxStack: operation.maxStack}
 		}
 		if !contract.UseGas(cost) {
 			return nil, ErrOutOfGas