--- a/core/vm/stack_table.go
+++ b/core/vm/stack_table.go
@@ -16,27 +16,14 @@
 
 package vm
 
-import (
-	"github.com/ethereum/go-ethereum/params"
-)
-
 func minSwapStack(n int) int {
 	return minStack(n, n)
-}
-func maxSwapStack(n int) int {
-	return maxStack(n, n)
 }
 
 func minDupStack(n int) int {
 	return minStack(n, n+1)
 }
-func maxDupStack(n int) int {
-	return maxStack(n, n+1)
-}
 
-func maxStack(pop, push int) int {
-	return int(params.StackLimit) + pop - push
-}
 func minStack(pops, push int) int {
 	return pops
 }