Fixes https://github.com/ethereum/go-ethereum/issues/24065