@gumb0 2021-12-07
master vs Martin's refactoring (first commit)
```
name                                  old time/op  new time/op  delta
main/blake2b_shifts/2805nulls         14.2ms ±22%  13.1ms ± 8%  -7.32%  (p=0.000 n=20+18)
main/blake2b_shifts/5610nulls         29.6ms ± 1%  29.3ms ± 1%  -1.05%  (p=0.000 n=18+19)
main/blake2b_shifts/8415nulls         43.9ms ± 1%  43.1ms ± 1%  -1.82%  (p=0.000 n=20+17)
main/weierstrudel/0                    471µs ± 0%   470µs ± 1%    ~     (p=0.159 n=18+19)
main/weierstrudel/1                    973µs ± 1%   969µs ± 0%  -0.38%  (p=0.000 n=17+20)
main/weierstrudel/3                   1.51ms ± 0%  1.51ms ± 1%  -0.17%  (p=0.012 n=20+19)
main/weierstrudel/9                   3.10ms ± 1%  3.10ms ± 0%    ~     (p=0.945 n=20+19)
main/weierstrudel/14                  4.42ms ± 0%  4.42ms ± 1%    ~     (p=0.196 n=19+17)
main/sha1_divs/empty                   195µs ± 1%   196µs ± 1%  +0.65%  (p=0.000 n=20+20)
main/sha1_divs/1351                   3.82ms ± 2%  3.87ms ± 1%  +1.55%  (p=0.000 n=18+19)
main/sha1_divs/2737                   7.35ms ± 1%  7.54ms ± 1%  +2.64%  (p=0.000 n=19+18)
main/sha1_divs/5311                   14.4ms ± 1%  14.8ms ± 1%  +2.82%  (p=0.000 n=20+20)
main/sha1_shifts/empty                 153µs ± 0%   155µs ± 0%  +1.40%  (p=0.000 n=16+18)
main/sha1_shifts/1351                 3.05ms ± 1%  3.08ms ± 1%  +1.03%  (p=0.000 n=19+19)
main/sha1_shifts/2737                 5.91ms ± 1%  5.98ms ± 1%  +1.12%  (p=0.000 n=17+19)
main/sha1_shifts/5311                 11.5ms ± 1%  11.8ms ± 1%  +1.83%  (p=0.000 n=19+20)
main/blake2b_huff/empty                101µs ± 1%   102µs ± 1%  +1.53%  (p=0.000 n=19+20)
main/blake2b_huff/2805nulls           1.91ms ± 1%  1.92ms ± 1%  +0.71%  (p=0.000 n=19+18)
main/blake2b_huff/5610nulls           3.67ms ± 1%  3.72ms ± 1%  +1.30%  (p=0.000 n=19+19)
main/blake2b_huff/8415nulls           5.39ms ± 2%  5.43ms ± 1%  +0.74%  (p=0.006 n=20+19)
micro/JUMPDEST_n0/empty               2.42ms ± 0%  2.43ms ± 0%  +0.46%  (p=0.000 n=19+20)
micro/jump_around/empty                140µs ± 1%   142µs ± 1%  +2.02%  (p=0.000 n=19+17)
micro/memory_grow_mstore/nogrow        433µs ± 1%   440µs ± 1%  +1.49%  (p=0.000 n=17+18)
micro/memory_grow_mstore/by1           460µs ± 1%   467µs ± 1%  +1.52%  (p=0.000 n=18+20)
micro/memory_grow_mstore/by16          575µs ± 6%   585µs ± 0%  +1.80%  (p=0.000 n=16+17)
micro/memory_grow_mstore/by32          663µs ± 0%   666µs ± 3%    ~     (p=0.479 n=16+20)
micro/signextend/zero                  310µs ± 1%   318µs ± 1%  +2.69%  (p=0.000 n=16+18)
micro/signextend/one                   319µs ± 1%   330µs ± 1%  +3.58%  (p=0.000 n=19+20)
micro/loop_with_many_jumpdests/empty  1.82µs ± 2%  1.83µs ± 3%  +0.98%  (p=0.009 n=20+19)
micro/memory_grow_mload/nogrow         212µs ± 2%   210µs ± 1%  -0.66%  (p=0.001 n=20+17)
micro/memory_grow_mload/by1            228µs ± 1%   227µs ± 1%  -0.42%  (p=0.034 n=18+19)
micro/memory_grow_mload/by16           315µs ± 2%   318µs ± 3%  +0.73%  (p=0.027 n=19+19)
micro/memory_grow_mload/by32           415µs ± 1%   412µs ± 6%    ~     (p=0.684 n=17+20)
[Geo mean]                            1.12ms       1.13ms       +0.59%
```