@gumb0 2021-12-07
master vs  optimization with `mayOverflow` flag
```
name                                  old time/op  new time/op  delta
main/blake2b_shifts/2805nulls         14.2ms ±22%  14.1ms ±11%     ~     (p=0.211 n=20+20)
main/blake2b_shifts/5610nulls         29.6ms ± 1%  30.6ms ± 1%   +3.31%  (p=0.000 n=18+20)
main/blake2b_shifts/8415nulls         43.9ms ± 1%  45.2ms ± 2%   +3.15%  (p=0.000 n=20+19)
main/weierstrudel/0                    471µs ± 0%   477µs ± 0%   +1.45%  (p=0.000 n=18+18)
main/weierstrudel/1                    973µs ± 1%   985µs ± 0%   +1.32%  (p=0.000 n=17+18)
main/weierstrudel/3                   1.51ms ± 0%  1.52ms ± 0%   +0.72%  (p=0.000 n=20+19)
main/weierstrudel/9                   3.10ms ± 1%  3.11ms ± 0%   +0.37%  (p=0.000 n=20+20)
main/weierstrudel/14                  4.42ms ± 0%  4.45ms ± 0%   +0.64%  (p=0.000 n=19+18)
main/sha1_divs/empty                   195µs ± 1%   203µs ± 1%   +3.88%  (p=0.000 n=20+19)
main/sha1_divs/1351                   3.82ms ± 2%  3.98ms ± 0%   +4.19%  (p=0.000 n=18+17)
main/sha1_divs/2737                   7.35ms ± 1%  7.79ms ± 2%   +5.94%  (p=0.000 n=19+18)
main/sha1_divs/5311                   14.4ms ± 1%  15.2ms ± 2%   +6.00%  (p=0.000 n=20+17)
main/sha1_shifts/empty                 153µs ± 0%   162µs ± 1%   +6.06%  (p=0.000 n=16+20)
main/sha1_shifts/1351                 3.05ms ± 1%  3.23ms ± 1%   +6.10%  (p=0.000 n=19+19)
main/sha1_shifts/2737                 5.91ms ± 1%  6.29ms ± 1%   +6.42%  (p=0.000 n=17+19)
main/sha1_shifts/5311                 11.5ms ± 1%  12.3ms ± 1%   +6.56%  (p=0.000 n=19+20)
main/blake2b_huff/empty                101µs ± 1%   104µs ± 1%   +3.19%  (p=0.000 n=19+19)
main/blake2b_huff/2805nulls           1.91ms ± 1%  1.96ms ± 1%   +2.48%  (p=0.000 n=19+20)
main/blake2b_huff/5610nulls           3.67ms ± 1%  3.78ms ± 1%   +2.89%  (p=0.000 n=19+20)
main/blake2b_huff/8415nulls           5.39ms ± 2%  5.52ms ± 0%   +2.49%  (p=0.000 n=20+19)
micro/JUMPDEST_n0/empty               2.42ms ± 0%  2.50ms ± 1%   +3.59%  (p=0.000 n=19+20)
micro/jump_around/empty                140µs ± 1%   145µs ± 3%   +3.76%  (p=0.000 n=19+20)
micro/memory_grow_mstore/nogrow        433µs ± 1%   433µs ± 2%     ~     (p=0.149 n=17+20)
micro/memory_grow_mstore/by1           460µs ± 1%   458µs ± 0%   -0.55%  (p=0.000 n=18+19)
micro/memory_grow_mstore/by16          575µs ± 6%   575µs ± 1%     ~     (p=0.297 n=16+18)
micro/memory_grow_mstore/by32          663µs ± 0%   655µs ± 4%   -1.14%  (p=0.001 n=16+18)
micro/signextend/zero                  310µs ± 1%   373µs ± 3%  +20.29%  (p=0.000 n=16+15)
micro/signextend/one                   319µs ± 1%   388µs ± 3%  +21.83%  (p=0.000 n=19+18)
micro/loop_with_many_jumpdests/empty  1.82µs ± 2%  1.84µs ± 2%   +1.48%  (p=0.000 n=20+18)
micro/memory_grow_mload/nogrow         212µs ± 2%   209µs ± 1%   -1.32%  (p=0.000 n=20+19)
micro/memory_grow_mload/by1            228µs ± 1%   225µs ± 1%   -1.00%  (p=0.000 n=18+19)
micro/memory_grow_mload/by16           315µs ± 2%   317µs ± 4%   +0.54%  (p=0.003 n=19+17)
micro/memory_grow_mload/by32           415µs ± 1%   414µs ± 8%     ~     (p=0.244 n=17+20)
[Geo mean]                            1.12ms       1.16ms        +3.32%
```