@holiman 2021-12-07
Seems odd. Here's what I get
```
$ benchstat master first 
name                    old time/op    new time/op    delta
SimpleLoop/loop-100M-6     456ms ±11%     418ms ± 3%   -8.34%  (p=0.003 n=10+9)

name                    old alloc/op   new alloc/op   delta
SimpleLoop/loop-100M-6    2.75kB ±12%    3.07kB ± 0%  +11.76%  (p=0.008 n=10+8)

name                    old allocs/op  new allocs/op  delta
SimpleLoop/loop-100M-6      24.8 ± 5%      26.0 ± 0%   +4.84%  (p=0.005 n=10+9)

```