@holiman 2021-12-07
For some reason, this was even faster
```
		if sLen := stack.len(); sLen < operation.minStack || sLen == int(params.StackLimit) {
			// todo figure out the _actual_ error
			return nil, ErrStackOverflow
		}

```
```
$ benchstat master foo
name                    old time/op    new time/op    delta
SimpleLoop/loop-100M-6     456ms ±11%     404ms ± 4%  -11.40%  (p=0.000 n=10+10)
```