@gumb0 2021-12-08
> For some reason, this was even faster

This has one check less (so doesn't check overflows correctly)