@gumb0 2021-12-21
> Marked as draft -- are you still working on this?

Yes, we plan at least to try benchmarking it on different hardware.
cc @chfast