@holiman 2022-11-11
> Still marked as draft -- untouched for 9 months. Shall we leave it open, revisit and open properly, or close?

Closing this for now, fee free to reopen if you pick it up again.